Rename `zip.hpp` to `Zip.hpp` for consistency.
