+++ include/h2/meta/TypeList.hpp
@@ -28,6 +28,6 @@
 #include "typelist/Size.hpp"
 #include "typelist/Sort.hpp"
 #include "typelist/Unique.hpp"
-#include "typelist/zip.hpp"
+#include "typelist/Zip.hpp"
 
 #endif // H2_META_TYPELIST_HPP_
