+++ include/h2/meta/typelist/CMakeLists.txt
@@ -34,5 +34,5 @@ h2_add_sources_to_target_and_install(
   Sort.hpp
   TypeList.hpp
   Unique.hpp
-  zip.hpp
+  Zip.hpp
   )
